requirements.txt
langchain
langchain-core
langchain-community
faiss-cpu
tiktoken
chromadb
sentence-transformers
python-dotenv
langchain-groq
pypdf
pymupdf
pdfminer
python-docx
docx2txt
unstructured
pandas
numpy
scikit-learn
openpyxl
msoffcrypto-tool
jq
matplotlib
seaborn
plotly
pyforest

## Embedding Packages
langchain-huggingface
sentence-transformers
langchain-openai